backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
openai==1.5.0
